user can click on edit trip and be taken to the start city map view
