+++ app/views/trips/show.html.erb
@@ -1,1 +1,3 @@
+<%= link_to "Edit Trip", edit_user_trip_path(@trip.user, @trip) %>
+
 <h3>Itinerary for <%= @trip.start_city %> on <%= @trip.start_date %> - <%= @trip.end_date %></h3>
